Updating and shifting around
